==> notebooks/Fig5/panel_A/infected_enrichment/4.NOC_processing.ipynb <==
# NOC processing
#### This is the fourth and final step of the enrichment calculation
Calculate NOC proportion and append to the enrichment table  

Key output files
- `yyyy-mm-dd_enrichment_table_NOC_prop.csv`: enrichment values and NOC proportion  
- `yyyy-mm-dd_volcano_table.csv`: enrichment values and p-values  
  
both files are in the directory `output/enrichment_and_volcano_tables`

import pandas as pd
import sys
import os
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np
from datetime import datetime
from pathlib import Path

script_path = Path.cwd().parent.parent.parent.parent / "script"
data_path = Path.cwd().parent.parent.parent.parent / "data"
sys.path.append(str(script_path))

from pyseus import basic_processing as ip
from pyseus import contrast_tools as ct

### Load imputed NOC/Bulk table
The correct datestamp is required to find the input files

timestamp = datetime.now().strftime('%Y-%m-%d')
print(f"Timestamp: {timestamp}")
outprefix = f"{timestamp}_QC_filter_impute"

outdir = Path.cwd() / "output"
preprocessing_out = outdir / f"preprocessing"

NOC_path = preprocessing_out / f"{outprefix}_filtered_table.csv"

try:
    noc_table = pd.read_csv(NOC_path, header=[0], index_col=0)
except FileNotFoundError:
    print(f"File {NOC_path} not found.\nPlease rerun the previous steps or specify the correct timestamp, current value is {timestamp}")
except pd.errors.ParserError:
    print(f"There was an error parsing the CSV file at {NOC_path}.")
except Exception as e:
    print(f"An unexpected error occurred: {e}")


### Calculation of proportion

# take only the noc columns 
NOC_kw = ["Cytosol", "Nuclear", "Organelle"]
noc_cols = [col for col in list(noc_table) if (
    any(kw in col for kw in NOC_kw)
) and 'Proteome' not in col]
print(noc_cols)

# just select the noc_cols + Protein IDs for merge purpose
noc_table = noc_table[['Protein IDs'] + noc_cols].copy()

# group the tables and take the median

# start RawTables class from basic processing to group the samples
noc_process = ip.RawTables(proteingroup=None, file_designated=True,
    info_cols=['Protein IDs'], sample_cols=noc_cols)

# designate the pre-computed noc filtered table on this class
noc_process.filtered_table = noc_table

# transform the intensity to linear space
#noc_process.transform_intensities(func=np.exp2)

# group replicates
noc_process.group_replicates(reg_exp=r'(.*)_\d+$')

# calculate the median
noc_medians = ip.median_replicates(noc_process.grouped_table)
noc_medians.head()

# we then take the sum of each column 
noc_medians['noc_sum_inf'] = noc_medians['Infected_Cytosol'] + noc_medians['Infected_Organelle'] + noc_medians['Infected_Nuclear']


# remove all proteingroups where both inf and unif noc_sum is 0 -> there is no information on these
noc_medians = noc_medians[(noc_medians['noc_sum_inf'] > 0) ].reset_index(drop=True)


# create a new dataframe with just the proportion: intensity divided by the sum
noc_proportion = noc_medians[['Protein IDs']].copy()
noc_proportion['NOC_cytosol_Infected'] = noc_medians['Infected_Cytosol'] / noc_medians['noc_sum_inf']
noc_proportion['NOC_organelle_Infected'] = noc_medians['Infected_Organelle'] / noc_medians['noc_sum_inf']
noc_proportion['NOC_nuclear_Infected'] = noc_medians['Infected_Nuclear'] / noc_medians['noc_sum_inf']


### Merge NOC proportion with IP enrichment

# finally, merge with the enrichment table

# read the enrichment table
enrich_out_dir = outdir / "enrichment_and_volcano_tables"
enrichment_csv_path = enrich_out_dir /  f'{timestamp}_enrichment_table.csv'
enrichments = pd.read_csv(enrichment_csv_path, header=[0,1], index_col=0)
meta_cols = enrichments["metadata"].columns
enrichments = enrichments.droplevel(0, axis=1)

# finally, merge with the enrichment table
enrichment_nocs = enrichments.merge(noc_proportion, on='Protein IDs', how='inner')

# standard formatting
enrichment_nocs = ct.standard_pyseus_headers(enrichment_nocs, meta_cols = meta_cols)

# save the enrichment table with NOC proportions
enrichment_NOC_csv_path = enrich_out_dir / f'{timestamp}_enrichment_table_NOC_prop.csv'
enrichment_nocs.to_csv(enrichment_NOC_csv_path)

# enrichment table with NOC proportions
enrichment_nocs
